+++ test/obsidian/path_spec.lua
@@ -121,6 +121,11 @@ describe("Path.with_suffix()", function()
       Path.new("/"):with_suffix ".png"
     end)
   end)
+
+  it("should allow appending the suffix instead of replacing it", function()
+    assert.equals(Path.new "johnny.md", Path.new("johnny.decimal"):with_suffix ".md")
+    assert.equals(Path.new "johnny.decimal.md", Path.new("johnny.decimal"):with_suffix(".md", true))
+  end)
 end)
 
 describe("Path.is_absolute()", function()